refactor(connectors): isolate Windows credential checks
