+++ src/main/connectors/custom-mcp-bootstrap.ts
@@ -5,6 +5,7 @@ import { ALL_CONNECTOR_IDS } from './registry'
 import { validateResourceId } from '../../shared/resource-id'
 import { customServerSecurityFingerprint } from '../settings/custom-server-identity'
 import { isSecureCustomMcpUrl } from './custom-mcp-url'
+import { hasAmbiguousCustomMcpCredentialNames } from './custom-mcp-windows-credential-names'
 
 export type CustomMcpFailureAvailability = 'unavailable' | 'unauthenticated'
 
@@ -55,22 +56,6 @@ const SUPPORTED_CUSTOM_MCP_TRANSPORTS = new Set<StoredCustomMcpServer['transport
   'streamable_http',
   'sse'
 ])
-
-const hasCaseInsensitiveNameCollision = (values: Record<string, string> | undefined): boolean => {
-  const names = Object.keys(values ?? {})
-  return new Set(names.map((name) => name.toLowerCase())).size !== names.length
-}
-
-export const hasAmbiguousCustomMcpCredentialNames = (
-  fields: Pick<StoredCustomMcpServer, 'transport' | 'env' | 'envRefs' | 'headers' | 'headerRefs'>,
-  platform = process.platform
-): boolean =>
-  fields.transport === 'stdio'
-    ? platform === 'win32' &&
-      (hasCaseInsensitiveNameCollision(fields.envRefs) ||
-        hasCaseInsensitiveNameCollision(fields.env))
-    : hasCaseInsensitiveNameCollision(fields.headerRefs) ||
-      hasCaseInsensitiveNameCollision(fields.headers)
 
 // A name already owned by a bundled Connector or another custom record remains visible in Settings
 // but cannot be exposed or dispatched.
